EVP_PKEY_bits instead of EVP_PKEY_size, require >=openssl-1.1.1
EVP_PKEY_size() is being used to get the key size in bytes, but it
doesn't do that. It returns a suggested crypto buffer size. To get the
key size instead use EVP_PKEY_bitrs() which returns the key size in
bits.

Also increase the minimum openssl version from 0.9.7 to 1.1.1.